+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Ability to change slides by mouse drag
 - `track-transition-duration` property
+- `no-drag` property
 - Documentation for browser support
 - Prettier
 - `aria-live="polite"` to track